fix: Updated docs and bugfixes

+++ source/Public/Merge-ADAuditZip.ps1
@@ -1,14 +1,12 @@
 function Merge-ADAuditZip {
-    <#
+        <#
     .SYNOPSIS
     Combines multiple audit report files into a single compressed ZIP file.
-
     .DESCRIPTION
     The Merge-ADAuditZip function combines multiple audit report files into a single
     compressed ZIP file. The function takes an array of file paths, a maximum file
     size for the output ZIP file, an output folder for the merged file, and an optional
     switch to open the directory of the merged file after creation.
-
     .PARAMETER FilePaths
     Specifies an array of file paths to be merged into a single compressed ZIP file.
     .PARAMETER MaxFileSize
@@ -39,14 +37,15 @@ function Merge-ADAuditZip {
     name, where X is the part number.
 
     This function may or may not work with various types of input.
-#>
+    #>
     param(
         [string[]]$FilePaths, # Array of file paths to be merged into a single zip file
         [int]$MaxFileSize = 24MB, # Maximum size (in bytes) of the output zip file
         [string]$OutputFolder = "C:\temp", # Output path of the merged zip file
-        [switch]$OpenDirectory  # Optional switch to open the directory of the merged zip file after creation
+        [switch]$OpenDirectory # Optional switch to open the directory of the merged zip file after creation
     )
-    [string]$OutputPath = "$($OutputFolder)\$((Get-Date).ToString('yyyy-MM-dd_hh.mm.ss'))_$($env:USERDNSDOMAIN)_CombinedAudit.zip"
+    # Remove any blank file paths from the array
+    $FilePaths = $FilePaths | Where-Object { $_ }
     # Create the output directory if it doesn't exist
     if (-not (Test-Path -Path $OutputFolder)) {
         $Script:ADLogString += Write-AuditLog -Message "Would you like to create the directory $($OutputFolder)?" -Severity Warning
@@ -62,28 +61,25 @@ function Merge-ADAuditZip {
     # Create a hashtable to store the file sizes
     $fileSizes = @{}
     foreach ($filePath in $FilePaths) {
-        # Check if the file is not empty
-        if (Get-Content $filePath) {
-            $fileSizes[$filePath] = (Get-Item $filePath).Length  # Get the size of each file and store in hashtable
-        }
-        else {
-            $Script:ADLogString += Write-AuditLog -Message  "File $filePath is empty and will be skipped."
-        }
+        $fileSizes[$filePath] = (Get-Item $filePath).Length  # Get the size of each file and store in hashtable
     }
+
     # Sort the files by size in descending order
     $sortedFiles = $fileSizes.GetEnumerator() | Sort-Object -Property Value -Descending | Select-Object -ExpandProperty Name
+
+    # Build the output path
+    $dateTimeString = (Get-Date).ToString('yyyy-MM-dd_hh.mm.ss')
+    $domainName = $env:USERDNSDOMAIN
+    $outputFileName = "$($dateTimeString)_$($domainName)_CombinedAudit.zip"
+    $outputPath = Join-Path $OutputFolder $outputFileName
     # Add files to the zip until the maximum size is reached
     $currentSize = 0
-    $partNumber = 1
     $filesToAdd = @()
     foreach ($filePath in $sortedFiles) {
         if (($currentSize + $fileSizes[$filePath]) -gt $MaxFileSize) {
             # If adding the next file would exceed the maximum size
             # Create a zip file with the current batch of files
-            $partName = "part$partNumber"
-            $partNumber++
-            $partPath = Join-Path $OutputFolder "$OutputFileBaseName-$partName.zip"
-            Compress-Archive -Path $filesToAdd -DestinationPath $partPath -Update
+            Compress-Archive -Path $filesToAdd -DestinationPath $outputPath -Update
             $filesToAdd = @()  # Clear the list of files to add
             $currentSize = 0  # Reset current size counter
         }
@@ -91,19 +87,18 @@ function Merge-ADAuditZip {
         $currentSize += $fileSizes[$filePath]  # Add the size of the current file to the current size counter
     }
     # Create a zip file with the remaining files
-    $partName = "part$partNumber"
-    $partPath = Join-Path $OutputFolder "$OutputFileBaseName-$partName.zip"
-    Compress-Archive -Path $filesToAdd -DestinationPath $partPath -Update
-
+    Compress-Archive -Path $filesToAdd -DestinationPath $outputPath -Update
     # Remove the original files
     foreach ($filePath in $FilePaths) {
-        Remove-Item -Path $filePath -Force
+        if ($filePath) {
+            Remove-Item -Path $filePath -Force
+        }
     }
     if ($OpenDirectory) {
         # If the OpenDirectory switch is used
-        Invoke-Item (Split-Path $OutputPath)  # Open the directory of the merged zip file
+        Invoke-Item (Split-Path $outputPath)  # Open the directory of the merged zip file
     }
     else {
-        return $OutputPath  # Otherwise, return the path of the merged zip file
+        return $outputPath  # Otherwise, return the path of the merged zip file
     }
 }
